Developer

+++ eGuide.AdminUI/src/app/components/facilities-components/facility-list/facility-list.component.ts
@@ -2,6 +2,7 @@ import { FacilityService } from 'src/app/services/facility.service';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Facility } from 'src/app/models/facility';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-facility-list',
@@ -33,13 +34,12 @@ export class FacilityListComponent implements OnInit {
 
   openForm() {
     this.isOpen = !this.isOpen;
-    console.log('open form');
+    this.facilityForm.reset();
   }
 
   getFacilities() {
     this.facilityService.getFacilities().subscribe((facilities) => {
       this.facilities = facilities;
-      console.log(this.facilities);
     });
   }
 
@@ -59,6 +59,7 @@ export class FacilityListComponent implements OnInit {
 
   toggleEdit(facility: Facility) {
     // other sockets should be disabled
+    this.isOpen = false;
     this.facilities.forEach((element) => {
       element.editingMode = false;
     });
@@ -67,32 +68,50 @@ export class FacilityListComponent implements OnInit {
 
   closeEdit(facility: Facility) {
     facility.editingMode = false;
+    this.facilityForm.reset();
   }
 
   editFacility(facility: Facility) {
     const facilityId = facility.id;
-    console.log(facilityId);
-    this.facilityService
-      .updateFacility(facilityId, this.facilityForm.value)
-      .subscribe(
-        (data) => {
-          console.log(data);
-          this.getFacilities();
-        },
-        (error) => {
-          console.log(error);
-        }
-      );
+    if (facilityId != null) {
+      this.facilityService
+        .updateFacility(facilityId, this.facilityForm.value)
+        .subscribe({
+          next: () => {
+            this.getFacilities();
+          },
+          error: (error) => {
+            console.log(error);
+          },
+        });
+    }
   }
 
-  deleteFacility(id: any) {
-    this.facilityService.deleteFacility(id).subscribe(
-      () => {
-        this.getFacilities();
-      },
-      (error) => {
-        console.log(error);
+  deleteFacility(id: string) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.facilityService.deleteFacility(id).subscribe({
+          next: () => {
+            this.getFacilities();
+            Swal.fire({
+              title: 'Deleted!',
+              text: 'Your file has been deleted.',
+              icon: 'success',
+            });
+          },
+          error: (error) => {
+            console.log(error);
+          },
+        });
       }
-    );
+    });
   }
 }
